+++ service/src/test/java/org/databiosphere/workspacedataservice/dao/MockBackupDaoConfig.java
@@ -14,6 +14,6 @@ public class MockBackupDaoConfig {
     @Profile("mock-backup-dao")
     @Primary
     BackupRestoreDao<BackupResponse> mockBackupDao() {
-        return new MockBackupRestoreDao(CloneTable.BACKUP);
+        return new MockBackupDao(CloneTable.BACKUP);
     }
 }
