+++ service/src/test/java/org/databiosphere/workspacedataservice/dao/MockRestoreDaoConfig.java
@@ -14,6 +14,6 @@ public class MockRestoreDaoConfig {
     @Profile("mock-restore-dao")
     @Primary
     BackupRestoreDao<RestoreResponse> mockRestoreDao() {
-        return new MockBackupRestoreDao(CloneTable.RESTORE);
+        return new MockBackupDao(CloneTable.RESTORE);
     }
 }
